pyproject.toml
[project]
name = "wmaug-appsync-talk"
version = "0.1.0"
description = "Example AWS AppSync Demo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.171.0",
    "boto3>=1.35.80",
    "cdk-klayers>=0.3.0",
    "polyfactory>=2.18.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.3",
]

[dependency-groups]
dev = []
